+++ libs/babobook/include/simple/simple_order.h
@@ -25,9 +25,9 @@ public:
   SimpleOrder();
 
   SimpleOrder(bool is_buy,
-              uint32_t price,
-              uint32_t qty,
-              uint32_t stop_price = 0,
+              uint64_t price,
+              uint64_t qty,
+              uint64_t stop_price = 0,
               book::OrderConditions conditions = book::OrderCondition::oc_no_conditions);
 
   /// @brief test hook: reset the auto-increment id generator so the next
@@ -45,42 +45,42 @@ public:
   bool is_buy() const;
 
   /// @brief get the limit price of this order
-  uint32_t price() const;
+  uint64_t price() const;
 
-  uint32_t stop_price()const;
+  uint64_t stop_price()const;
 
   /// @brief get the quantity of this order
-  uint32_t order_qty() const;
+  uint64_t order_qty() const;
 
   /// @brief get the open quantity of this order (remaining, minus any reserved)
-  uint32_t open_qty() const;
+  uint64_t open_qty() const;
 
   /// @brief tentatively set aside (reserve) quantity during AON deferred-match
   /// planning, without filling it. Lowers the reported open_qty() so the same
   /// quantity isn't double-counted while an AON match is being assembled.
   /// @param reserved amount to add to the reservation (may be negative to release)
   /// @return the remaining available quantity after this reservation
-  uint32_t reserve(int32_t reserved);
+  uint64_t reserve(int64_t reserved);
 
   /// @brief is there no remaining open quantity in this order?
   bool filled() const;
 
   /// @brief modify the order's quantity by delta (used by replace).
   /// Mirrors the old OrderTracker::change_qty now that the order tracks itself.
-  void change_qty(int32_t delta);
+  void change_qty(int64_t delta);
 
   /// @brief get the filled quantity of this order
-  const uint32_t& filled_qty() const;
+  const uint64_t& filled_qty() const;
 
   /// @brief get the total filled cost of this order
-  const uint32_t& filled_cost() const;
+  const uint64_t& filled_cost() const;
 
   /// @brief notify of a fill of this order
   /// @param fill_qty the number of shares in this fill
   /// @param fill_cost the total amount of this fill
   /// @fill_id the unique identifier of this fill
-  void fill(uint32_t fill_qty,
-                    uint32_t fill_cost,
+  void fill(uint64_t fill_qty,
+                    uint64_t fill_cost,
                     uint32_t fill_id);
 
   /// @brief get order conditions as a bit mask
@@ -103,23 +103,23 @@ public:
   /// @brief exchange replaced this order
   /// @param size_delta change to the order quantity
   /// @param new_price the new price
-  void replace(uint32_t size_delta, uint32_t new_price);
+  void replace(uint64_t size_delta, uint64_t new_price);
 
   /// @brief unconditionally modify qty/price for a book replace (keeps id + filled history).
   /// @param size_delta change to the total order quantity (may be negative)
   /// @param new_price the new limit price, or PRICE_UNCHANGED (0) to keep the current price
-  void modify(int32_t size_delta, uint64_t new_price);
+  void modify(int64_t size_delta, uint64_t new_price);
 
 private:
   OrderState state_;
   bool is_buy_;
-  uint32_t order_qty_;
-  uint32_t price_;
-  uint32_t stop_price_;
+  uint64_t order_qty_;
+  uint64_t price_;
+  uint64_t stop_price_;
   book::OrderConditions conditions_;
-  uint32_t filled_qty_;
-  uint32_t filled_cost_;
-  int32_t reserved_;            // quantity tentatively earmarked during AON planning (not yet filled)
+  uint64_t filled_qty_;
+  uint64_t filled_cost_;
+  int64_t reserved_;            // quantity tentatively earmarked during AON planning (not yet filled)
   inline static uint32_t last_order_id_{0};
 
 public:
